requirements.txt
asgiref==3.2.10
django-admin-easy==0.6
django-cacheops==5.0.1
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.3.0
django-import-export==2.3.0
Django==3.1.1
djangorestframework==3.12.1
drf-yasg==1.17.1
Markdown==3.2.2
psycopg2-binary==2.8.6
python-decouple==3.3
python-slugify==4.0.1
pytz==2020.1
sqlparse==0.3.1
